# feat: P0 Phase - High-performance IPC foundation with sub-100µs latency