--- a/cmd/pyproc/main.go
+++ b/cmd/pyproc/main.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"text/template"
 
 	"github.com/spf13/cobra"