--- a/cmd/pyproc/templates/go.mod.tmpl
+++ b/cmd/pyproc/templates/go.mod.tmpl
@@ -1,0 +1,7 @@
+module {{.GoModule}}
+
+go 1.21
+
+require (
+	github.com/YuminosukeSato/pyproc v0.1.0
+)