--- a/cmd/pyproc/templates/requirements.txt.tmpl
+++ b/cmd/pyproc/templates/requirements.txt.tmpl
@@ -1,0 +1,3 @@
+pyproc-worker>=0.1.0
+orjson>=3.9.0
+msgspec>=0.18.0  # Optional: for high-performance codecs