… .github/workflows/ci.yml
… .golangci.yml
… .serena/cache/go/document_symbols_cache_v23-06-25.pkl
… .serena/memories/project_overview.md
… .serena/memories/tech_stack_conventions.md
… .serena/project.yml
… api/v1/pyproc.pb.go
… api/v1/pyproc_grpc.pb.go
… bench/benchmark_test.go
… bench/python_servers/base_server.py
… bench/python_servers/jsonrpc_server.py
… bench/python_servers/msgpack_server.py
… bench/python_servers/xmlrpc_server.py
… bench/rpc_clients/interface.go
… bench/rpc_clients/jsonrpc_client.go
… bench/rpc_clients/msgpack_client.go
… bench/rpc_clients/pyproc_client.go
… bench/rpc_clients/xmlrpc_client.go
… bench/rpc_comparison_benchmark_test.go
… docs/codec-performance.md
… examples/basic/main.go
… examples/basic/worker.py
… go.mod
… go.sum
… internal/framing/frame.go
… internal/framing/frame_test.go
… internal/framing/framing.go
… pkg/pyproc/codec.go
… pkg/pyproc/codec_json_goccy.go
… pkg/pyproc/codec_json_segmentio.go